A100/hqbbzsp/nips_submission_A100/submission_of_A100/submission 3/requirements.txt
torch==2.0.1
accelerate
appdirs
loralib
bitsandbytes
black
black[jupyter]
datasets
dataclasses
fire
peft
huggingface_hub
git+https://github.com/huggingface/transformers.git
sentencepiece
py7zr
scipy
optimum
transformers_stream_generator==0.0.4
tiktoken
einops
deepspeed